+++ jsonargparse_tests/typehints_tests.py
@@ -734,9 +734,9 @@ class TypeHintsTests(unittest.TestCase):
 
         with mock_module(MyCalendar) as module:
             parser = ArgumentParser()
-            parser.add_argument('--g1.op1', default=123)
+            parser.add_argument('--g1.op1', default=1)
             parser.add_argument('--g1.op2', default='abc')
-            parser.add_argument('--g2.op1', default=4.5)
+            parser.add_argument('--g2.op1', type=Callable, default=deepcopy)
             parser.add_argument('--g2.op2', type=Calendar, default=lazy_instance(Calendar, firstweekday=2))
 
             cfg = parser.get_defaults()
@@ -745,11 +745,17 @@ class TypeHintsTests(unittest.TestCase):
 
             cfg.g2.op2.class_path = f'{module}.MyCalendar'
             dump = parser.dump(cfg, skip_default=True)
-            self.assertEqual(dump, 'g2:\n  op2:\n    class_path: jsonargparse_tests.MyCalendar\n    init_args:\n      firstweekday: 2\n')
+            self.assertEqual(dump, f'g2:\n  op2:\n    class_path: {module}.MyCalendar\n    init_args:\n      firstweekday: 2\n')
 
             cfg.g2.op2.init_args.firstweekday = 0
             dump = parser.dump(cfg, skip_default=True)
-            self.assertEqual(dump, 'g2:\n  op2:\n    class_path: jsonargparse_tests.MyCalendar\n')
+            self.assertEqual(dump, f'g2:\n  op2:\n    class_path: {module}.MyCalendar\n')
+
+            parser.link_arguments('g1.op1', 'g2.op2.init_args.firstweekday')
+            parser.link_arguments('g1.op2', 'g2.op2.init_args.param')
+            del cfg['g2.op2.init_args']
+            dump = parser.dump(cfg, skip_default=True)
+            self.assertEqual(dump, f'g2:\n  op2:\n    class_path: {module}.MyCalendar\n')
 
 
 class TypeHintsTmpdirTests(TempDirTestCase):