+++ CHANGELOG.rst
@@ -14,6 +14,16 @@ The semantic versioning only considers the public API as described in
 paths are considered internals and can change in any moment.
 
 
+v4.9.1 (2022-06-??)
+-------------------
+
+Fixed
+^^^^^
+- ``dump`` with ``skip_default=True`` not working for subclasses without
+  ``init_args`` and when a default value requires serializing.
+- ``JSONARGPARSE_DEFAULT_ENV`` should have precedence over given value.
+
+
 v4.9.0 (2022-06-01)
 -------------------